arrumando maluquices
