--- src/pages/EnrolledList.tsx
@@ -1,10 +1,0 @@
-// import { useState } from "react";
-import TeacherView from "./EditalDetails/TeacherView";
-
-export default function EnrolledList() {
-  return (
-    <div>
-      <TeacherView />
-    </div>
-  );
-}
