+++ src/routes/ProtectedRoutes.tsx
@@ -1,5 +1,5 @@
 import { RouteProps } from "react-router-dom";
-import EnrolledList from "../pages/EnrolledList";
+import TeacherView from "../pages/EditalDetails/TeacherView";
 import Inscricao from "../pages/Inscricao/View";
 import auth from "../services/Auth";
 import { routesWithRedirect } from "./RoutesHelper";
@@ -14,7 +14,7 @@ TODO: Futuramente, desmembrar entre:
 
 const protectedRoutes: RouteProps[] = [
   { path: "/inscricao-edital/:editalId", element: <Inscricao /> },
-  { path: "/:editalId/inscritos", element: <EnrolledList /> },
+  { path: "/:editalId/inscritos", element: <TeacherView /> },
 
   // Exemplos:
   // { path: "/estudante", element: <Home /> },
